Deprecate the {File|Image}Uploader{Field} components
