+++ src/components/_DeprecatedFileUploader/index.stories.js
