+++ src/components/_DeprecatedImageUploader/index.stories.js
