+++ src/components/_DeprecatedImageUploaderField/index.stories.js
