+++ src/components/index.js
@@ -5,16 +5,12 @@ export { default as Checkbox } from './Checkbox';
 export { default as Dropdown } from './Dropdown';
 export { default as Datepicker } from './Datepicker';
 export { default as FormField } from './FormField';
-export { default as FileUploader } from './FileUploader';
-export { default as FileUploaderField } from './FileUploaderField';
 export { default as Grid } from './Grid';
 export { default as GridCol } from './Grid/GridCol';
 export { default as Heading } from './Heading';
 export { default as Helper } from './Helper';
 export { default as Icon } from './Icon';
 export { default as IconButton } from './IconButton';
-export { default as ImageUploader } from './ImageUploader';
-export { default as ImageUploaderField } from './ImageUploaderField';
 export { default as Input } from './Input';
 export { default as InputField } from './InputField';
 export { default as Item } from './Item';
@@ -41,3 +37,9 @@ export { default as Selector } from './Selector';
 export { default as SelectorField } from './SelectorField';
 export { default as FormElement } from './FormComponents/FormElement';
 export { default as FormElementWrapper } from './FormComponents/FormElementWrapper';
+
+// deprecated
+export { default as FileUploader } from './_DeprecatedFileUploader';
+export { default as FileUploaderField } from './_DeprecatedFileUploaderField';
+export { default as ImageUploader } from './_DeprecatedImageUploader';
+export { default as ImageUploaderField } from './_DeprecatedImageUploaderField';
